Se añade el precio de la ruta y se corrige el campo del model voucher

+++ backend/proyectoDjango/aplicacionComprobanteTransporte/models/voucher.py
@@ -5,7 +5,7 @@ from .ruta import Ruta
 class Voucher(models.Model):
     cliente = models.ForeignKey(Cliente, on_delete=models.CASCADE, related_name='vouchers')
     ruta = models.ForeignKey(Ruta, on_delete=models.CASCADE)
-    cost = models.IntegerField(null=True, blank=True)
+    cost = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)  
 
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
